requirements.txt
chatbotAI
wikipedia
pyjokes
requests
flask
gradio